+++ CHANGELOG.md
@@ -17,6 +17,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 - Clean up and rename a large number of heuristic and filter functions.
 - Changed the default score function of the `Mapper` from `equality` to `disabled`.
+- The `TranslatorFactory` now makes an effort to include the source file of config issues.
 
 ### Fixed
 - Duplicate explicit names are now supported for most types (closes #4).